LayoutTests/imported/w3c/web-platform-tests/html/semantics/scripting-1/the-script-element/module/import-meta/import-meta-resolve-multiple-scripts-expected.txt


PASS import.meta.resolve resolves URLs relative to the import.meta.url, not relative to the active script when it is called: another global's inline script
PASS import.meta.resolve still works if its global has been destroyed (by detaching the iframe)
PASS import.meta.resolve resolves URLs relative to the import.meta.url, not relative to the active script when it is called: another module script
